Refactor dashboard page and extract hooks
This commit refactors the dashboard page by extracting several functionalities into dedicated hooks:
- `useDashboardNotifications`: Manages upvote notifications, including fetching, real-time updates via WebSockets, and read status.
- `useDashboardPublishing`: Handles the portfolio publishing logic, including validation, API calls, and user feedback.
- `useDashboardSummary`: Fetches and displays community update summaries (new upvotes).
- `useDashboardFeedPrefetch`: Implements prefetching for the feed section to improve load times.
- `useDashboardResume`: Manages the logic for generating and downloading ATS resumes.

Additionally, a `DashboardSectionRenderer` component is introduced to simplify the rendering of different dashboard sections, improving code organization and maintainability. The `DashboardLayout` now utilizes these new hooks and components.
